Inlined some singleton functions, broke out XHR into its own method
