+++ form.js
@@ -2,12 +2,10 @@
 import log from "./logging"
 import { is_array, is_function, duplicate } from "./utils"
 
-import { modal, update_hidden_fields} from "./visuals" //commented-out: tooltip
+import { modal} from "./visuals"
 import { tooltip } from "./tooltip"
-import validator from "./validator";
-
-// import { domainToUnicode } from "url"
-// import { stringify } from "querystring"
+import DnsValidator from "./dns_validator";
+import xhr from "./xhr.js"
 
 /************
  * The intention here is that we won't put any GUI stuff in here at all.
@@ -58,7 +56,7 @@ export default class Form {
             this.timeout = 1000 * this.timeout //timeout was in seconds, but window.setTimeout() is in milliseconds
         }
 
-        this.doh_server = new validator(this.doh_json_server,this.timeout)
+        this.doh_server = new DnsValidator(this.doh_json_server,this.timeout)
 
         if(this.manual) {
             //bail out of the rest of setup for manual-mode
@@ -277,6 +275,28 @@ export default class Form {
         }
     }
 
+    update_hidden_fields(checksum, status) {
+        this.insert_or_update_hidden('goodforms_checksum',checksum)
+        this.insert_or_update_hidden('goodforms_status',status)
+    }
+
+    insert_or_update_hidden(name,value) {
+        var possibles = this.form.elements //getElementsByName(name)
+        for (let possible in possibles) {
+            if(possible.name == name) {
+                possible.value = value
+                return
+            }
+        }
+        var new_elem = document.createElement('input')
+        let attributes = {'type': 'hidden','name': name,'value': value}
+        for (let key in attributes) {
+            new_elem.setAttribute(key,attributes[key])
+        }
+        this.form.appendChild(new_elem)
+    }
+
+
     // Event-management/hook-management helper methods
 
     parse_event_handler_results(result, behavior, visuals, allow_func, params) {
@@ -349,15 +369,16 @@ export default class Form {
     }
 
     ongood_handler(detailed_status, checksum, message) {
+        log.verbose("Uh, detailed status is, uh: "+detailed_status)
         this.fire_hooks('onGood', () => {
             if(this.form) {
-                update_hidden_fields(this.form, checksum, status) //TODO: why is this struckthrough?
+                this.update_hidden_fields(checksum, detailed_status)
             }
             this.enable_submits()
         },
         () => {
             if(this.visuals.good) {
-                this.tooltip.hide()
+                this.tooltip.hide() // TODO - is this right?
             }    
         },
         detailed_status,
@@ -411,7 +432,7 @@ export default class Form {
         })
     } 
 
-     onerror_handler() {
+    onerror_handler() {
         this.fire_hooks('onError',() => {
             log.debug("Error detected?")
             this.tooltip.remove()
@@ -454,82 +475,8 @@ export default class Form {
             this.doh_server.verify(data, complete) // TODO - pass timeout?
         } else {
             //do server-side validation via GoodForms
-            let req = null
-            if(typeof XMLHttpRequest !== "undefined") {
-                req = new XMLHttpRequest()
-            } else if(window.ActiveXObject) {
-                log.verbose("trying old-school way");
-                try {
-                    req = new ActiveXObject("Msxml2.XMLHTTP")
-                } catch (e) {
-                    log.verbose("trying the other weird old way?")
-                    try {
-                        req = new ActiveXObject("Microsoft.XMLHTTP")
-                    } catch (e) {
-                        //silently 'eat' error
-                        log.error("no XMLHttpRequest available...");
-                    }
-                }
-                log.verbose("Setting timeouts");
-                if(typeof req.setTimeouts !== "undefined") {
-                    log.verbose("Setting timeotus for really reals!")
-                    req.setTimeouts(this.timeout,this.timeout,this.timeout,this.timeout)
-                } else {
-                    log.verbose("No timeouts we can set :(")
-                }
-                log.verbose("Timeouts set!");
-                // log.verbose("URL we're going to try to set is: "+window.location.href)
-                // req.setRequestHeader("origin",window.location.href)
-
-                req.onreadystatechange = () => {
-                    if(req.readyState == 4) {
-                        log.verbose("Ready state is up!")
-                        log.verbose("Actual state of the thing is: "+req.status)
-                        log.verbose("Status text? "+req.statusText)
-                        let raw_results = req.responseText
-                        log.verbose("Response text?"+raw_results)
-                        try {
-                            var results = JSON.parse(raw_results)
-                            log.verbose("Parsed Result: "+results)
-                        } catch (error) {
-                            log.verbose("Actual text: "+req.responseText)
-                            complete({status: "ERROR", message: "Invalid JSON response"})
-                            return
-                        }
-                        complete(results)
-                    }
-                }
-            }
-            req.open("POST", HOST+'/'+url,true)
-            req.setRequestHeader('accept', 'application/json')
-            req.setRequestHeader('content-type','application/json')
-            if(typeof req.timeout !== "undefined") {
-                req.timeout = this.timeout
-            }
-
-            if(typeof req.addEventListener !== "undefined") {
-                req.addEventListener('error',function (event) {
-                    complete({status: "ERROR", message: "Server Error"})
-                })
-                req.addEventListener('timeout', function () {
-                    complete({status: "ERROR", message: "Timeout"})
-                })
-                req.addEventListener('abort',function () {
-                    complete({status: "ERROR", message: "Aborted"})
-                })
-
-                req.addEventListener('load', function (event) {
-                    try {
-                        var results = JSON.parse(req.responseText)
-                    } catch (error) {
-                        complete({status: "ERROR", message: "Invalid JSON response"})
-                        return
-                    }
-                    complete(results)
-                })
-            }
-            log.verbose("Stringified data is: "+JSON.stringify(data))
-            req.send(JSON.stringify(data))
+            // yeah, harder than it seems - move the xhr to the inside of the resolver
+            xhr(HOST+'/'+url, data,complete)
         }
 
     }
